Implement new form components and new hooks structure - wip
