+++ src/app/Form/FormScores/FormScores.md
@@ -1,0 +1,5 @@
+Form Scores
+
+```jsx
+<FormScores />
+```
