+++ src/app/Form/FormScores/index.js
@@ -1,0 +1,1 @@
+export { default } from './FormScores';
